I discovered itertools!
I was being forced to calculate and use the tensor components as arrays outside of the Tensor class, then flatten them into 1 dimensional lists, and only then was I able to assign them to the instance attribute 'components' (now named 'use').  The reason for this was that in order to iterate through all of the components of a tensor/array (necessary for printing the components), I would need nested for loops, the number of which would vary depending on the rank of the tensor.  I didn't know of any way to write a function that could emulate a varying number of nested for loops until I came across itertools.  So now I can calculate and use the tensors' instance attribute directly as an array, plus I'm now able to use way fewer lines of code by compressing a bunch of nested for loops into one or two for loops.